use model library as base
